backend/pyproject.toml
[project]
name = "airesume"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "docx2txt>=0.9",
    "fastapi>=0.120.0",
    "langchain>=1.0.2",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.0.0",
    "langchain-huggingface>=1.0.0",
    "pdfminer-six>=20250506",
    "pip>=25.3",
    "pyresparser>=1.0.6",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.20",
    "spacy>=3.8.7",
    "uvicorn>=0.38.0",
]
